+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+- Fixed popup closing when visit status changes
+
 ## [0.10.2] - 2025-03-01
 ### Fixed
 - Repeated marker renderings on page load